@@ -1,7 +1,7 @@
 {
-    "name": "qlimix/http-diactoros",
+    "name": "qlimix/http-laminas",
     "type": "library",
-    "description": "Http helper interfaces implemented with diactoros and httphandlerrunner",
+    "description": "Http helper interfaces implemented with Laminas diactoros and httphandlerrunner",
     "license": "MIT",
     "keywords": [
         "http"
@@ -13,11 +13,14 @@
         }
     ],
     "require": {
-        "php": ">=7.2",
+        "php": ">7.4",
         "psr/http-server-handler": "^1.0",
-        "qlimix/http": "^1.0",
-        "zendframework/zend-diactoros": "^2.1",
-        "zendframework/zend-httphandlerrunner": "^1.1"
+        "qlimix/http": "^2.0",
+        "laminas/laminas-diactoros": "^2.5",
+        "laminas/laminas-httphandlerrunner": "^1.3"
+    },
+    "require-dev": {
+        "qlimix/code-standard": "^3.0"
     },
     "autoload": {
         "psr-4": {
@@ -31,8 +34,5 @@
     },
     "config": {
         "sort-packages": true
-    },
-    "require-dev": {
-        "qlimix/code-standard": "^1.0"
     }
 }
